+++ assets/svelte/components/PropertiesSidebar.svelte
@@ -1,12 +1,21 @@
 <script lang="ts">
+  import { createEventDispatcher } from "svelte"
   import Pill from "$lib/components/Pill.svelte"
   import SidebarSection from "$lib/components/SidebarSection.svelte"
-  import { createEventDispatcher } from "svelte"
-  import { draggedObject } from "$lib/stores/dragAndDrop"
+  import { draggedComponentDefinition } from "$lib/stores/dragAndDrop"
   import { live } from "$lib/stores/live"
-  import { updateNodeContent } from "$lib/utils/ast-manipulation"
-  import { page, selectedAstElement, selectedAstElementId, findAstElement, isAstElement } from "$lib/stores/page"
+  import {
+    page,
+    selectedAstElement,
+    selectedAstElementId,
+    findAstElement,
+    isAstElement,
+    setSelection,
+    resetSelection,
+  } from "$lib/stores/page"
   import type { AstNode } from "$lib/types"
+  import { getParentNodeId } from "$lib/utils/ast-helpers"
+  import { deleteAstNode, updateNodeContent } from "$lib/utils/ast-manipulation"
   import { elementCanBeDroppedInTarget } from "$lib/utils/drag-helpers"
 
   const dispatch = createEventDispatcher()
@@ -59,18 +68,9 @@
     }
   }
 
-  function parentNodeId() {
-    if ($selectedAstElementId) {
-      let parts = $selectedAstElementId.split(".")
-      if (parts.length === 1) return "root"
-      return parts.slice(0, -1).join(".")
-    }
-  }
   function selectParentNode() {
-    let parentId = parentNodeId()
-    if (parentId) {
-      $selectedAstElementId = parentId
-    }
+    let parentId = getParentNodeId($selectedAstElementId)
+    setSelection(parentId)
   }
 
   async function deleteClass(className: string) {
@@ -106,17 +106,11 @@
   }
 
   async function deleteComponent() {
-    let node = $selectedAstElement
-    if (!node) return
+    if (!$selectedAstElementId) return
+
     if (confirm("Are you sure you want to delete this component?")) {
-      let parentId = parentNodeId()
-      let content = parentId && parentId !== "root" ? findAstElement($page.ast, parentId)?.content : $page.ast
-      if (content) {
-        let targetIndex = (content as unknown[]).indexOf(node)
-        content.splice(targetIndex, 1)
-        $selectedAstElementId = undefined
-        $live.pushEvent("update_page_ast", { id: $page.id, ast: $page.ast })
-      }
+      deleteAstNode($selectedAstElementId)
+      resetSelection()
     }
   }
 
@@ -174,7 +168,7 @@
             </svg>
           </button>
         {/if}
-        <button type="button" class="absolute p-2 top-2 right-1" on:click={() => ($selectedAstElementId = undefined)}>
+        <button type="button" class="absolute p-2 top-2 right-1" on:click={resetSelection}>
           <span class="sr-only">Close</span>
           <svg
             xmlns="http://www.w3.org/2000/svg"
@@ -202,7 +196,6 @@
         {#each editableAttrs as entry (entry)}
           {@const [name, value] = entry}
           <SidebarSection
-            clearOnUpdate={true}
             {value}
             on:delete={() => deleteAttribute(name)}
             on:textChange={(e) => updateAttribute(name, e)}
@@ -238,11 +231,17 @@
         </div>
       {/if}
       {#if $selectedAstElement.tag === "eex_block"}
-        <SidebarSection on:update={updateArg} value={$selectedAstElement.arg} large={true}>
+        <SidebarSection
+          on:update={updateArg}
+          disabled={true}
+          value={$selectedAstElement.arg}
+          large={true}
+          disableDelete={true}
+        >
           <svelte:fragment slot="heading">Block argument</svelte:fragment>
           <svelte:fragment slot="input"></svelte:fragment>
         </SidebarSection>
-        <SidebarSection>
+        <SidebarSection disableDelete={true}>
           <svelte:fragment slot="heading">Block content</svelte:fragment>
           <svelte:fragment slot="input">
             <p>The content of eex blocks can't be edited from the visual editor yet. Please use the code editor.</p>
@@ -251,7 +250,7 @@
       {/if}
 
       <div class="relative">
-        {#if $draggedObject && elementCanBeDroppedInTarget($draggedObject)}
+        {#if $draggedComponentDefinition && elementCanBeDroppedInTarget($draggedComponentDefinition)}
           <div
             class="absolute bg-white opacity-70 w-full h-full p-4"
             class:opacity-90={isDraggingOver}