+++ assets/svelte/stores/page.ts
@@ -1,22 +1,22 @@
 import { writable, derived, get } from "svelte/store"
 import type { Writable, Readable } from "svelte/store"
 import type { AstElement, AstNode, Page } from "$lib/types"
-import type { CoordsDiff } from "$lib/utils/drag-helpers"
 
 export const page: Writable<Page> = writable()
 export const selectedAstElementId: Writable<string | undefined> = writable()
-// export const highlightedAstElementId: Writable<string | undefined> = writable();
 export const highlightedAstElement: Writable<AstElement | undefined> = writable()
 export const slotTargetElement: Writable<AstElement | undefined> = writable()
 
 export const rootAstElement: Readable<AstElement | undefined> = derived([page], ([$page]) => {
   // This is a virtual AstElement intended to simulate the page itself to reorder the components at the first level.
-  return { tag: "root", attrs: {}, content: $page.ast }
+  if ($page) {
+    return { tag: "root", attrs: {}, content: $page.ast }
+  }
 })
 export const selectedAstElement: Readable<AstElement | undefined> = derived(
   [page, selectedAstElementId],
   ([$page, $selectedAstElementId]) => {
-    if ($selectedAstElementId) {
+    if ($page && $selectedAstElementId) {
       if ($selectedAstElementId === "root") return get(rootAstElement)
       return findAstElement($page.ast, $selectedAstElementId)
     }
@@ -26,7 +26,7 @@ export const selectedAstElement: Readable<AstElement | undefined> = derived(
 export const parentOfSelectedAstElement: Readable<AstElement | undefined> = derived(
   [page, selectedAstElementId],
   ([$page, $selectedAstElementId]) => {
-    if ($selectedAstElementId) {
+    if ($page && $selectedAstElementId) {
       if ($selectedAstElementId === "root") return null
       let levels = $selectedAstElementId.split(".")
       if (levels.length === 1) return get(rootAstElement)
@@ -37,16 +37,19 @@ export const parentOfSelectedAstElement: Readable<AstElement | undefined> = deri
 )
 
 export const selectedDomElement: Writable<Element | null> = writable(null)
-export interface SelectedElementMenu {
-  top: number
-  left: number
-  dragging: boolean
-  dragDirection: "horizontal" | "vertical"
-  elementCoords: CoordsDiff
-  insertBefore: number | null
-  mouseMovement: CoordsDiff
+
+export function setSelection(selectedId: string) {
+  selectedAstElementId.update(() => selectedId)
 }
-export const selectedElementMenu: Writable<SelectedElementMenu | null> = writable(null)
+
+export function setSelectedDom(selectedDom: Element) {
+  selectedDomElement.update(() => selectedDom)
+}
+
+export function resetSelection() {
+  selectedAstElementId.update(() => null)
+  selectedDomElement.update(() => null)
+}
 
 export function isAstElement(maybeNode: AstNode): maybeNode is AstElement {
   return typeof maybeNode !== "string"
@@ -79,4 +82,12 @@ export function _findAstElementId(ast: AstNode[], astNode: AstNode, id: string):
       }
     }
   }
+}
+
+export function resetStores() {
+  page.set(null)
+  selectedAstElementId.set(null)
+  highlightedAstElement.set(null)
+  slotTargetElement.set(null)
+  selectedDomElement.set(null)
 }