+++ lib/beacon/live_admin/components/admin_components.ex
@@ -10,7 +10,7 @@ defmodule Beacon.LiveAdmin.AdminComponents do
   alias Beacon.LiveAdmin.CoreComponents
   alias Beacon.LiveAdmin.PageBuilder.Table
 
-  import Beacon.LiveAdmin.Gettext
+  use Gettext, backend: Beacon.LiveAdmin.Gettext
   import Beacon.LiveAdmin.Router, only: [beacon_live_admin_path: 3]
 
   defdelegate header(assigns), to: CoreComponents
@@ -141,9 +141,6 @@ defmodule Beacon.LiveAdmin.AdminComponents do
     </li>
     <li class="mr-2">
       <.link patch={beacon_live_admin_path(@socket, @site, "/pages/#{@page_id}/variants")} class={if(@current_action == :variants, do: @active_class, else: @regular_class)}>Variants</.link>
-    </li>
-    <li class="mr-2">
-      <.link patch={beacon_live_admin_path(@socket, @site, "/pages/#{@page_id}/events")} class={if(@current_action == :events, do: @active_class, else: @regular_class)}>Events</.link>
     </li>
     <li class="mr-2">
       <.link patch={beacon_live_admin_path(@socket, @site, "/pages/#{@page_id}/revisions")} class={if(@current_action == :revisions, do: @active_class, else: @regular_class)}>Revisions</.link>