+++ lib/beacon/live_admin/gettext.ex
@@ -1,5 +1,5 @@
 defmodule Beacon.LiveAdmin.Gettext do
   @moduledoc false
 
-  use Gettext, otp_app: :beacon_live_admin
+  use Gettext.Backend, otp_app: :beacon_live_admin
 end