+++ CHANGELOG.md
@@ -1,14 +1,34 @@
 # Changelog
 
+## 0.1.0-dev
+
+### Breaking Changes
+  * Require minimum Elixir v1.14
+  * Require minimum Gettext v0.26 to use the new backend module
+
+### Fixes
+  - Close #226 - Fix a race condition between update AST and save change events #240
+  - Close #204 - skip `PageBuilder.Table.handle_params/3` when the requested page has no pagination data
+
+### Enhancements
+- [Visual Editor] Improve DnD highlight states and simplify logic significantly ([#219](https://github.com/BeaconCMS/beacon_live_admin/pull/219))
+- [Visual Editor] Better detect overlapping when dragging elements to reorder ([#216](https://github.com/BeaconCMS/beacon_live_admin/pull/216))
+- [Visual Editor] Display a delete icon on selected elements ([#209](https://github.com/BeaconCMS/beacon_live_admin/pull/209))
+- [Visual Editor] Better detect horizontal/vertical drag and drop flow ([#215](https://github.com/BeaconCMS/beacon_live_admin/pull/215))
+
+### Fixes
+- [Dev] Fix tailwind watch config
+- [Visual Editor] Fix drag button orientation ([#218](https://github.com/BeaconCMS/beacon_live_admin/pull/218))
+- [Visual Editor] Do not show drag buttons on elements that are only children ([#217](https://github.com/BeaconCMS/beacon_live_admin/pull/217))
+- [Visual Editor] Keep current element select after drag and drop event ([#214](https://github.com/BeaconCMS/beacon_live_admin/pull/214))
+
 ## 0.1.0-rc.1 (2024-08-27)
 
 ### Enhancements
-
 - [Event Handler] Added Event Handlers ([#195](https://github.com/BeaconCMS/beacon_live_admin/pull/195))
 - [Visual Editor] Allow to reorder an element among its siblings with drag and drop ([#174](https://github.com/BeaconCMS/beacon_live_admin/pull/174))
 
 ### Fixes
-
 - [Visual Editor] Disable dragLeave trigger on drag placeholder ([#208](https://github.com/BeaconCMS/beacon_live_admin/pull/208))
 - [Visual Editor] Reset drag states when dropping, even on invalid targets ([#206](https://github.com/BeaconCMS/beacon_live_admin/pull/206))
 - [Visual Editor] Disable interacting with iframes ([#198](https://github.com/BeaconCMS/beacon_live_admin/pull/198))
@@ -17,7 +37,6 @@
 ## 0.1.0-rc.0 (2024-08-02)
 
 ### Enhancements
-
 - Media Library
   - Upload, preview, and display image location
   - Sorting and pagination